Refactoring: Calendar에서 날짜형식 formatDate.ts로 분리
