+++ src/components/Detail/Calendar/index.tsx
@@ -11,19 +11,11 @@ import {
   checkInDateState,
   checkOutDateState,
 } from "../../../store/checkInCheckOutAtom.ts";
+import formatDate from "../../../utils/formatDate";
 
 const Calendar: React.FC<CalendarProps> = ({ price }) => {
   const [checkInDate, setCheckInDate] = useRecoilState(checkInDateState);
   const [checkOutDate, setCheckOutDate] = useRecoilState(checkOutDateState);
-
-  // 날짜를 로컬시간대로 변환하고, yyyy-mm-dd 형식으로 변환.
-  const formatDate = (date: Date | null) => {
-    if (!date) return "";
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    return `${year}-${month}-${day}`;
-  };
 
   // 날짜 변경 핸들러
   const handleChange = (dates: [Date, Date]) => {
